Change display toggle box to div
The display toggle button now is a div inside of a div (instead of a div
in a input checkbox). This solved the issues of clicking on the checkbox
label instead of the circle causing the button to move but not change
the display. Also labels were added to show the display options.